Modifie les variables en float

+++ minitrice.py
@@ -23,8 +23,8 @@ def calculator(input_str):
 
 
     try:
-        left_num = int(left)
-        right_num = int(right)
+        left_num = float(left)
+        right_num = float(right)
     except ValueError:
         print("Erreur de syntaxe pour le calcul: " + input_str)
         return
